analytics/requirements.txt
setuptools>=68.0.0
wheel>=0.41.0
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
numpy>=1.26.0
scipy>=1.11.0
pandas>=2.1.0
openpyxl>=3.1.2
gunicorn>=21.2.0
